+++ led_and_button.h
@@ -2,6 +2,6 @@
 #include "PORTS_INIT.h"
 #include "E:/keil/EE319Kware/inc/tm4c123gh6pm.h"
 void led_button_init (void) ;
-void led1_on (uint8_t distance) ;
+void led1_on (float distance) ;
 uint8_t get_button1_value(void); //start point
 uint8_t get_button2_value(void); //end point
